# services/banking-integration-service/src/main/java/com/manh/partnerbridge/banking/domain/model/BankAccount.java
package com.manh.partnerbridge.banking.domain.model;

import java.math.BigDecimal;

public record BankAccount(String accountId, String holderName, BigDecimal balance, String currency, String status) {}