# services/banking-integration-service/src/main/java/com/manh/partnerbridge/banking/infrastructure/error/ApiError.java
package com.manh.partnerbridge.banking.infrastructure.error;

import java.time.Instant;
import java.util.List;

public record ApiError(String code, String message, String traceId, String requestId,
                       Instant timestamp, List<ApiErrorDetail> details) { }